[WPF][Comptabilite] Assembleur WPF + CodeBehind + Maj Ref WCF

+++ SupErp/SupERP.WPF.Comptabiity/Model/TransactionWPF.cs
@@ -9,12 +9,12 @@ namespace SupERP.WPF.Comptabiity.Model
     public class TransactionWPF
     {
         public String libelle { get; set; }
-        public float montant { get; set; }
+        public decimal montant { get; set; }
         public DateTime date { get; set; }
         public bool paye { get; set; }
         public String client { get; set; }
 
-        public TransactionWPF(String _libelle, float _montant, DateTime _date, bool _paye, String _client)
+        public TransactionWPF(String _libelle, decimal _montant, DateTime _date, bool _paye, String _client)
         {
             libelle = _libelle;
             montant = _montant;
